+++ src/index.ts
@@ -1,7 +1,7 @@
-import Select from './Select';
+import Select, { SelectProps } from './Select';
 import Option from './Option';
 import OptGroup from './OptGroup';
 
-export { Option, OptGroup };
+export { Option, OptGroup, SelectProps };
 
 export default Select;
